navigation-api/navigate-event/defaultPrevented-window-stop-after-dispatch.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  let navigateEvent;
  navigation.onnavigate = e => navigateEvent = e;
  const navigationToStop = navigation.navigate("?1");
  assert_false(navigateEvent.defaultPrevented);

  window.stop();
  await assertBothRejectDOM(t, navigationToStop, "AbortError");
  assert_false(navigateEvent.defaultPrevented);
}, "window.stop() doesn't affect navigateEvent.defaultPrevented after dispatch");
</script>
